+++ examples/RigidBodyCases/RodRigidBodyContact/rod_cylinder_contact_validation.py
@@ -5,7 +5,7 @@ import sys
 
 sys.path.append("../../../")
 from elastica import *
-from matplotlib import pyplot as plt
+from post_processing import plot_video, plot_cylinder_rod_position
 
 
 class SingleRodSingleCylinderInteractionSimulator(
@@ -30,6 +30,7 @@ action_plane_key = "x"
 
 # can be set to True, checks collision at tips of rod
 TIP_COLLISION = True
+TIP_CHOICE = 1
 
 _roll_key = 0 if action_plane_key == "x" else (1 if action_plane_key == "y" else 2)
 if action_plane_key == "x":
@@ -71,8 +72,8 @@ cylinder_radius = 10.0 * base_radius
 tip_offset = 0.0
 if TIP_COLLISION:
     # The random choice decides which tip of the rod intersects with cylinder
-    tip_choice = np.random.choice([1, -1])
-    tip_offset = 0.5 * tip_choice * base_length * np.cos(inclination)
+    TIP_CHOICE = np.random.choice([1, -1])
+    tip_offset = 0.5 * TIP_CHOICE * base_length * np.cos(inclination)
 
 start_rod_1 = np.array(
     [
@@ -153,135 +154,21 @@ print("Total steps", total_steps)
 integrate(timestepper, single_rod_sim, final_time, total_steps)
 
 if PLOT_FIGURE:
-
-    def make_data_for_cylinder_along_z(cstart, cradius, cheight):
-        center_x, center_y = cstart[0], cstart[1]
-        z = np.linspace(0, cheight, 5)
-        theta = np.linspace(0, 2 * np.pi, 20)
-        theta_grid, z_grid = np.meshgrid(theta, z)
-        x_grid = cradius * np.cos(theta_grid) + center_x
-        y_grid = cradius * np.sin(theta_grid) + center_y
-        z_grid += cstart[2]
-        return np.roll([x_grid, y_grid, z_grid], _roll_key)
-
-    XC, YC, ZC = make_data_for_cylinder_along_z(
-        cylinder_start, cylinder_radius, cylinder_height
+    plot_video(
+        recorded_rod_history,
+        recorded_cyl_history,
+        "cylinder_rod_collision.mp4",
+        cylinder_direction=cylinder_direction,
+        cylinder_height=cylinder_height,
+        cylinder_radius=cylinder_radius,
     )
 
-    def plot_video(
-        plot_params: dict,
-        cylinder_history: dict,
-        video_name="video.mp4",
-        margin=0.2,
-        fps=60,
-        step=1,
-        *args,
-        **kwargs
-    ):  # (time step, x/y/z, node)
-        import matplotlib.animation as manimation
-
-        plt.rcParams.update({"font.size": 22})
-
-        # Should give a (n_time, 3, n_elem) array
-        positions = np.array(plot_params["position"])
-        # (n_time, 3) array
-        com = np.array(plot_params["com"])
-
-        cylinder_com = np.array(cylinder_history["com"])
-        cylinder_origin = cylinder_com - 0.5 * cylinder_height * cylinder_direction
-
-        print("plot video")
-        FFMpegWriter = manimation.writers["ffmpeg"]
-        metadata = dict(
-            title="Movie Test", artist="Matplotlib", comment="Movie support!"
-        )
-        writer = FFMpegWriter(fps=fps, metadata=metadata)
-        dpi = 50
-
-        # min_limits = np.roll(np.array([0.0, -0.5 * cylinder_height, 0.0]), _roll_key)
-        from mpl_toolkits.mplot3d import Axes3D
-
-        fig = plt.figure(1, figsize=(10, 8), frameon=True, dpi=dpi)
-        ax = plt.axes(projection="3d")  # fig.add_subplot(111)
-        ax.grid(b=True, which="minor", color="k", linestyle="--")
-        ax.grid(b=True, which="major", color="k", linestyle="-")
-        # plt.axis("square")
-        i = 0
-        (rod_line,) = ax.plot(positions[i, 0], positions[i, 1], positions[i, 2], lw=3.0)
-        XC, YC, ZC = make_data_for_cylinder_along_z(
-            cylinder_origin[i, ...], cylinder_radius, cylinder_height
-        )
-        surf = ax.plot_surface(XC, YC, ZC, color="g", alpha=0.5)
-        ax.set_xlabel("x")
-        ax.set_ylabel("y")
-        ax.set_zlabel("z")
-
-        min_limits = global_rot_mat @ np.array([0.0, -0.5 * cylinder_height, 0.0])
-        min_limits = -np.abs(min_limits)
-        max_limits = min_limits + cylinder_height
-
-        ax.set_xlim([min_limits[0], max_limits[0]])
-        ax.set_ylim([min_limits[1], max_limits[1]])
-        ax.set_zlim([min_limits[2], max_limits[2]])
-        with writer.saving(fig, video_name, dpi):
-            with plt.style.context("seaborn-white"):
-                for i in range(0, positions.shape[0], int(step)):
-                    rod_line.set_xdata(positions[i, 0])
-                    rod_line.set_ydata(positions[i, 1])
-                    rod_line.set_3d_properties(positions[i, 2])
-                    XC, YC, ZC = make_data_for_cylinder_along_z(
-                        cylinder_origin[i, ...], cylinder_radius, cylinder_height
-                    )
-                    surf.remove()
-                    surf = ax.plot_surface(XC, YC, ZC, color="g", alpha=0.5)
-                    writer.grab_frame()
-        from matplotlib.patches import Circle
-
-        fig = plt.figure(2, figsize=(10, 8), frameon=True, dpi=dpi)
-        ax = fig.add_subplot(111)
-        i = 0
-        positions = np.roll(positions, -_roll_key, axis=1)
-        com = np.roll(com, -_roll_key, axis=1)
-        cstart = np.roll(cylinder_origin, -_roll_key, axis=1)
-        (rod_line,) = ax.plot(positions[i, 0], positions[i, 1], lw=3.0)
-        (tip_line,) = ax.plot(com[i, 0], com[i, 1], "k--")
-
-        min_limits = np.array([0.0, -0.5 * cylinder_height, 0.0])
-        max_limits = min_limits + cylinder_height
-
-        ax.set_xlim([min_limits[0], max_limits[0]])
-        ax.set_ylim([min_limits[1], max_limits[1]])
-
-        circle_artist = Circle((cstart[i, 0], cstart[i, 1]), cylinder_radius, color="g")
-        ax.add_artist(circle_artist)
-        ax.set_aspect("equal")
-        video_name = "2D_" + video_name
-        with writer.saving(fig, video_name, dpi):
-            with plt.style.context("fivethirtyeight"):
-                for i in range(0, positions.shape[0], int(step)):
-                    rod_line.set_xdata(positions[i, 0])
-                    rod_line.set_ydata(positions[i, 1])
-                    tip_line.set_xdata(com[:i, 0])
-                    tip_line.set_ydata(com[:i, 1])
-                    circle_artist.center = cstart[i, 0], cstart[i, 1]
-                    writer.grab_frame()
-
-    plot_video(recorded_rod_history, recorded_cyl_history, "cylinder_rod_collision.mp4")
-
-    positions = np.array(recorded_rod_history["position"])
-    sim_time = np.array(recorded_rod_history["time"])
-    colliding_element_idx = n_elem // 2
-    if TIP_COLLISION:
-        colliding_element_idx = 0 if tip_choice == 1 else -1
-    colliding_element_history = positions[:, :, colliding_element_idx]
-    fig = plt.figure(3, figsize=(8, 5))
-    ax = fig.add_subplot(111)
-    ax.plot(sim_time, colliding_element_history[:, _roll_key])
-    ax.hlines(
-        cylinder_start[_roll_key] - cylinder_radius - base_radius,
-        sim_time[0],
-        sim_time[-1],
-        "k",
-        linestyle="dashed",
+    plot_cylinder_rod_position(
+        recorded_rod_history,
+        recorded_cyl_history,
+        cylinder_radius=cylinder_radius,
+        rod_base_radius=base_radius,
+        TIP_COLLISION=TIP_COLLISION,
+        TIP_CHOICE=TIP_CHOICE,
+        _roll_key=_roll_key,
     )
-    plt.show()
